+++ src/geophires_x_client/geophires_x_result.py
@@ -305,6 +305,9 @@ class GeophiresXResult:
                 "Drilling and completion costs (for redrilling)",
                 "Drilling and completion costs per redrilled well",
                 "Stimulation costs",
+                "Stimulation costs per well",
+                "Stimulation costs per production well",
+                "Stimulation costs per injection well",
                 "Stimulation costs (for redrilling)",
                 "Surface power plant costs",
                 "of which Electrical Plant Cost",
@@ -318,16 +321,17 @@ class GeophiresXResult:
                 "District Heating System Cost",
                 "Field gathering system costs",
                 "Total surface equipment costs",
+                "TESS capital costs",
                 "Investment Tax Credit",
                 "One-time Flat License Fees Etc",
-                "Other Incentives",
-                "One-time Grants Etc",
                 "Total Add-on CAPEX",
                 "Overnight Capital Cost",
                 # Displayed for economic models that treat inflation costs as capital costs (SAM-EM)
                 "Inflation costs during construction",
                 "Royalty supplemental payments during construction",
                 "Interest during construction",
+                "Other Incentives",
+                "One-time Grants Etc",
                 "Total capital costs",
                 "Annualized capital costs",
                 # AGS/CLGS
@@ -348,6 +352,7 @@ class GeophiresXResult:
                 "Water costs",
                 "Average Reservoir Pumping Cost",
                 "Absorption Chiller O&M Cost",
+                "TESS annual O&M costs",
                 "Average Heat Pump Electricity Cost",
                 "Annual District Heating O&M Cost",
                 "Average Annual Peaking Fuel Cost",
@@ -1115,6 +1120,7 @@ class GeophiresXResult:
                 "S-DAC-GT Annual Cost (USD/yr)",
                 "S-DAC-GT Cumm. Cash Flow (USD)",
                 "Cumm. Cost Per Tonne (USD/tonne)",
+                "Annual Carbon Revenue (USD/yr)",
             ]
 
         try: